Change package of new extensions module
`viewx`
