+++ view-extensions/src/androidTest/java/com/backbase/deferredresources/viewx/ViewTestActivity.kt
@@ -1,4 +1,4 @@
-package com.backbase.deferredresources.extensions
+package com.backbase.deferredresources.viewx
 
 import android.app.Activity
 import android.view.View
