+++ view-extensions/src/androidTest/java/com/backbase/deferredresources/viewx/ViewTestUtils.kt
@@ -1,4 +1,4 @@
-package com.backbase.deferredresources.extensions
+package com.backbase.deferredresources.viewx
 
 import android.content.Context
 import android.view.View
