+++ view-extensions/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest
     xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.backbase.deferredresources.extensions" />
+    package="com.backbase.deferredresources.viewx" />
